+++ OggConverter/src/Forms/Form1.cs
@@ -1084,6 +1084,14 @@ namespace OggConverter
 #endif
         private void Form1_KeyDown(object sender, KeyEventArgs e)
         {
+            // Refresh the localization (translator mode)
+            if (Settings.TranslatorMode)
+                if (e.KeyCode == Keys.F5)
+                {
+                    Localisation.LoadLocaleFile();
+                    Localize();
+                }
+
             // If the songlist is not focused and user presses up or down arrow - it will focus on song list and select the first song
             if ((e.KeyCode == Keys.Up || e.KeyCode == Keys.Down) && songList.SelectedIndex == -1 && this.ActiveControl != selectedFolder)
             {