+++ OggConverter/src/Forms/SettingsForm.Designer.cs
@@ -74,6 +74,7 @@
             this.chkHistory = new System.Windows.Forms.CheckBox();
             this.chkCrashLog = new System.Windows.Forms.CheckBox();
             this.labTranslator = new System.Windows.Forms.Label();
+            this.chkTranslatorMode = new System.Windows.Forms.CheckBox();
             this.tabControl.SuspendLayout();
             this.tabGeneral.SuspendLayout();
             this.tabFiles.SuspendLayout();
@@ -96,6 +97,7 @@
             // 
             // tabGeneral
             // 
+            this.tabGeneral.Controls.Add(this.chkTranslatorMode);
             this.tabGeneral.Controls.Add(this.labTranslator);
             this.tabGeneral.Controls.Add(this.btnAudacity);
             this.tabGeneral.Controls.Add(this.labAudacity);
@@ -115,7 +117,7 @@
             // btnAudacity
             // 
             this.btnAudacity.Font = new System.Drawing.Font("Microsoft Sans Serif", 12F);
-            this.btnAudacity.Location = new System.Drawing.Point(455, 298);
+            this.btnAudacity.Location = new System.Drawing.Point(455, 324);
             this.btnAudacity.Name = "btnAudacity";
             this.btnAudacity.Size = new System.Drawing.Size(44, 44);
             this.btnAudacity.TabIndex = 31;
@@ -126,7 +128,7 @@
             // labAudacity
             // 
             this.labAudacity.AutoSize = true;
-            this.labAudacity.Location = new System.Drawing.Point(27, 289);
+            this.labAudacity.Location = new System.Drawing.Point(27, 315);
             this.labAudacity.Name = "labAudacity";
             this.labAudacity.Size = new System.Drawing.Size(139, 17);
             this.labAudacity.TabIndex = 30;
@@ -134,7 +136,7 @@
             // 
             // txtAudacity
             // 
-            this.txtAudacity.Location = new System.Drawing.Point(29, 309);
+            this.txtAudacity.Location = new System.Drawing.Point(29, 335);
             this.txtAudacity.Name = "txtAudacity";
             this.txtAudacity.ReadOnly = true;
             this.txtAudacity.Size = new System.Drawing.Size(420, 22);
@@ -143,7 +145,7 @@
             // chkShowFfmpegOutput
             // 
             this.chkShowFfmpegOutput.AutoSize = true;
-            this.chkShowFfmpegOutput.Location = new System.Drawing.Point(29, 226);
+            this.chkShowFfmpegOutput.Location = new System.Drawing.Point(29, 252);
             this.chkShowFfmpegOutput.Name = "chkShowFfmpegOutput";
             this.chkShowFfmpegOutput.Size = new System.Drawing.Size(155, 21);
             this.chkShowFfmpegOutput.TabIndex = 28;
@@ -582,9 +584,20 @@
             this.labTranslator.AutoSize = true;
             this.labTranslator.Location = new System.Drawing.Point(26, 171);
             this.labTranslator.Name = "labTranslator";
-            this.labTranslator.Size = new System.Drawing.Size(99, 17);
+            this.labTranslator.Size = new System.Drawing.Size(77, 34);
             this.labTranslator.TabIndex = 32;
-            this.labTranslator.Text = "Translator: {0}";
+            this.labTranslator.Text = "Translator:\r\n{0}";
+            // 
+            // chkTranslatorMode
+            // 
+            this.chkTranslatorMode.AutoSize = true;
+            this.chkTranslatorMode.Location = new System.Drawing.Point(29, 225);
+            this.chkTranslatorMode.Name = "chkTranslatorMode";
+            this.chkTranslatorMode.Size = new System.Drawing.Size(134, 21);
+            this.chkTranslatorMode.TabIndex = 33;
+            this.chkTranslatorMode.Text = "Translator Mode";
+            this.chkTranslatorMode.UseVisualStyleBackColor = true;
+            this.chkTranslatorMode.Click += new System.EventHandler(this.chkTranslatorMode_Click);
             // 
             // SettingsForm
             // 
@@ -597,6 +610,7 @@
             this.MaximizeBox = false;
             this.Name = "SettingsForm";
             this.Text = "Settings";
+            this.KeyDown += new System.Windows.Forms.KeyEventHandler(this.SettingsForm_KeyDown);
             this.tabControl.ResumeLayout(false);
             this.tabGeneral.ResumeLayout(false);
             this.tabGeneral.PerformLayout();
@@ -657,5 +671,6 @@
         private System.Windows.Forms.CheckBox chkMono;
         private System.Windows.Forms.Label label7;
         private System.Windows.Forms.Label labTranslator;
+        private System.Windows.Forms.CheckBox chkTranslatorMode;
     }
 }